rm some waring in linux

Former-commit-id: 8ffffb556635098c2f0ae255233f06d0a6adceb6

+++ src/VmScript/TestMcu.cpp
@@ -5088,7 +5088,7 @@ string CTestMcu::MOVC_A_DPTR_Test(int space) {
 string CTestMcu::SUBB_A_RnTest(int space) {
 	byte code;
 	int PC = 0, aa = 0, rr = 0;
-	byte expect;
+	int expect;
 	byte cc;
 
 	auto checkresult = [&]()
